refactor(spec): split EAGLEWorker into BaseSpecWorker/BaseDraftWorker + EagleDraftWorker (P1-2)
Extract abstract base classes (BaseSpecWorker, BaseDraftWorker) and move
draft logic into EagleDraftWorker so MultiLayerEAGLEWorker/MultiLayerDraftWorker
(P1-4) can reuse the same verify/data-contract path.
